Allows the evil-state to be referenced directly from spaceline-install
Rationale:
vim-powerline (the spacemacs layer) allows you to see what state you are
in visually (as a word). It looks something like this:
NORMAL > other stuff

This is well aligned with what vim does by default, and makes it very
easy to see what state you are in, even for states you haven't
encountered before. For spacemacs, there are a lot more states than
normal, insert, and visual.

As it was before this commit, you cannot use evil-state anywhere in your
spaceline, since it will simply refer to the previously defined
evil-state segment. I have tried
```lisp
(spaceline-install

  '(((evil-state window-number)
     :fallback evil-state
     :face highlight-face)
     anzu
  ;; etc from the default spacemacs settings
```

That didn't work, for the reasons mentioned above. I've renamed the
segment to evil-current-state. I'm open to other suggestions for that
name, it was the first thing I came up with.