Don’t try to trim nil in battery status
fancy-battery mode asynchronously queries the battery status so the
status is unknown, i.e. nil, during early startup before the update time
ran for the first time.  In this case spaceline--fancy-battery-mode-line
returns nil but the battery segment unconditionally applies s-trim to
the return value, causing a wrong-type-argument during redisplay in
early startup.

To avoid this issue check whether any status exists and return nil from
the segment if not.

Closes GH-115