Add new depth argument to spaceline--generate
This is necessary because the previous DEEP argument was being used
for two things.

Previously, DEEP was true for top-level segments or segments that are
fallbacks for other top-level segments (and which should behave as if
they are top-level segments, i.e. generate separators).

Whenever DEEP was true, the length of the segment was also
computed. However, this should only be done for true top-level
segments, not fallbacks.

Therefore, introduce a new argument DEEP-OR-FALLBACK that takes the
place of the original DEEP (with a more descriptive name), and
re-fashion DEEP to have exact meaning of a segment being not
top-level.

Then, spaceline--gen-segment should use DEEP-OR-FALLBACK in all cases
where it previously used DEEP (except for the length computation), and
it should always, without reservation, pass a truthy value for the
DEEP argument when recursing.