Don't avoid rendering hidden segments
This always renders all segments, in order to compute their length,
but then resets the result pointer if the segment is hidden.

This allows spaceline to "discover" when segments that have previously
been hidden have become small enough to be shown again, even when the
cause for this is not that the window size has changed, but rather
that the segment itself simply is shorter (for whatever reason).

Details:

- Add a HIDDEN argument to spaceline--gen-segment. This should be a
  form to evaluate to check whether the segment is hidden. It is only
  useful for top-level segments.

- Separate the cleanup code into two parts: code that modifies RESULT
  and code that does not. The reason for this is that code that
  modifies RESULT must be executed for every segment that produces
  ouput, even hidden ones, since this may affect their
  length. However, code that does not modify RESULT must under no
  circumstances be executed for hidden segments.

- Clean up code is now the sole responsibility of the top-level
  segment, not the fallbacks. This has the effect of simplifying the
  code generation somewhat.