Remove overwrite of toggle vars value when defining segments
This commit privileges end users use cases instead of developers use cases when
it comes to Spaceline configuration. Explanation below:

This setq prevents easy setup of toggle variables when the mode line is lazy
loaded.

The natural way for a spacemacs user to disable a segment is to set the variable
or use the toggle functions. It should always work.

For spaceline developers creating new segments they should be aware of the
toggle variables and set them manually for the current session.