Added a simplified spaceline segment for version control
This version of the `version-control` segment does not describe what
VCS system you're use or what branch you're working on, which is not
useful information for people that mostly work on a Git master branch
anyway. It also minimizes the VCS information to a single character:
+ (addition), - (removal), ~ (editing), ! (needs action), ? (other).

The result fits neatly right after the `buffer-id` in the modeline,
as a VCS replacement for the standard `buffer-modified` section.