spaceline-info-mode: Attempt to load info+ if available
If info+ is available in the load-path but just doesn't happen to have
been loaded yet, spaceline-info-mode should load it.  This prevents
the confusing message:

    info+ is not available: spaceline-info-mode disabled

when info+ is in fact available.  Even worse, if info+ is loaded
immediately afterwards (and before the previous commit switching away
from defadvice), spaceline-info-mode is disabled but its effect is
still present.