spaceline-info-mode: Don't re-enable advice if info+ is loaded later
When spaceline-info-mode is enabled and the advice is run when info+
is not available, it should disable the advice.  However, the old
defadvice mechanism re-activates advice when the underlying function
is redefined.  This results in the confusing situation where
spaceline-info-mode is nil but the advice is enabled (and the
spaceline mode line is in effect).

This commit switches over to the new-style advice combinators, so that
spaceline-info-mode properly disables itself.